Assign incoming reports to buckets if possible via the quick methods.
This should speed up fresh and ongoing imports. It ignores reports that would match more than one bucket, which is a state we currently don't have (but could have in the future). That's fine, because the triage_new_reports cronjob will pick those up and then run the slower matching logic.